refactor: rename Traits to Concerns for better organization and clarity
