+++ app/Concerns/AuthorizationChecker.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Traits;
+namespace App\Concerns;
 
 use App\Models\User;
 use Illuminate\Contracts\Auth\Authenticatable;
