+++ app/Http/Controllers/Controller.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
-use App\Traits\AuthorizationChecker;
-use App\Traits\HasActionLogTrait;
+use App\Concerns\AuthorizationChecker;
+use App\Concerns\HasActionLogTrait;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
