+++ app/Models/Permission.php
@@ -2,7 +2,7 @@
 
 namespace App\Models;
 
-use App\Traits\QueryBuilderTrait;
+use App\Concerns\QueryBuilderTrait;
 use Spatie\Permission\Models\Permission as SpatiePermission;
 
 class Permission extends SpatiePermission
