+++ app/Services/EnvWriter.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Services;
 
 use App\Enums\ActionType;
-use App\Traits\HasActionLogTrait;
+use App\Concerns\HasActionLogTrait;
 
 class EnvWriter
 {
