+++ app/Services/SlugService.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Services;
 
-use App\Traits\HasUniqueSlug;
+use App\Concerns\HasUniqueSlug;
 
 class SlugService
 {
